--- src/chat-artmedia/main/java/ge/chat/artmedia/utils/ExtentReportManager.java
@@ -1,45 +1,0 @@
-package ge.chat.artmedia.utils;
-
-import com.aventstack.extentreports.ExtentReports;
-import com.aventstack.extentreports.ExtentTest;
-import com.aventstack.extentreports.reporter.ExtentSparkReporter;
-
-public class ExtentReportManager {
-    private static ExtentReports extent;
-    private static ThreadLocal<ExtentTest> test = new ThreadLocal<>();
-
-    //Initialize ExtentReports
-    public static ExtentReports getExtentReports() {
-        if (extent == null) {
-            String reportPath = System.getProperty("user.dir") + "/test-output/ExtentReport.html";
-            ExtentSparkReporter sparkReporter = new ExtentSparkReporter(reportPath);
-            sparkReporter.config().setReportName("Automation Test Report");
-            sparkReporter.config().setDocumentTitle("Test Execution Report");
-
-            extent = new ExtentReports();
-            extent.attachReporter(sparkReporter);
-            extent.setSystemInfo("Environment", "QA");
-            extent.setSystemInfo("Tester", "Your Name");
-        }
-        return extent;
-    }
-
-    //Create a test
-    public static ExtentTest createTest(String testName) {
-        ExtentTest extentTest = getExtentReports().createTest(testName);
-        test.set(extentTest);
-        return extentTest;
-    }
-
-    //Get the current test
-    public static ExtentTest getTest() {
-        return test.get();
-    }
-
-    //Flush reports
-    public static void flushReports() {
-        if (extent != null) {
-            extent.flush();
-        }
-    }
-}
