--- src/chat-artmedia/main/resources/testdata.properties
@@ -1,9 +1,0 @@
-crm.admin.email = liziberiashvili98@gmail.com
-crm.admin.wrong.email = wrongemail@gmail.com
-crm.admin.password = Test1234@
-crm.admin.wrong.password = TestTest
-
-
-
-crm.agent.password = Test1234!
-
