--- src/chat-artmedia/test/java/ge/chat/artmedia/tests/ui/DepartmentTests.java
@@ -1,4 +1,0 @@
-package ge.chat.artmedia.tests.ui;
-
-public class DepartmentTests{
-}
