--- src/chat-artmedia/test/resources/testdata.json
@@ -1,15 +1,0 @@
-{
-  "createFaq": {
-    "language": "en",
-    "question": "Question?",
-    "answer": "Answer",
-    "isVisible": true
-  },
-  "updateFaq": {
-    "id": "{faqId}",
-    "isVisible": true,
-    "language": "en",
-    "question": "Updated question?",
-    "answer": "Updated answer."
-  }
-}
