+++ src/main/java/ge/chat/artmedia/utils/PropertiesReader.java
