+++ src/main/java/ge/chat/artmedia/utils/TestDataFactory.java
@@ -1,0 +1,19 @@
+package ge.chat.artmedia.utils;
+
+import net.datafaker.Faker;
+
+public class TestDataFactory {
+    static Faker faker = new Faker();
+
+    public static String randomName(){
+        return faker.name().firstName();
+    }
+
+    public static String randomEmail() {
+        return faker.internet().emailAddress();
+    }
+
+    public static String randomMessage() {
+        return faker.lorem().sentence();
+    }
+}
